TaskService basic validations

+++ src/infrastructure/services/taskService.ts
@@ -30,10 +30,12 @@ export class TaskService {
     }
 
     async updateTask(task: Task): Promise<void> {
+        if (!task || task.id.length === 0) throw new Error("[Task Service - UpdateTask]: Cannot update task without task / task id"); 
         await this.taskRepository.update(task);
     }
 
     async deleteTask(id: string): Promise<void> {
+        if (!id) throw new Error("[Task Service - DeleteTask]: Task id cannot be empty");
         await this.taskRepository.delete(id);
     }
